Folder renamed from customTimeRanges to custom_time_ranges. Minor changes in editor.html. Validation for range.newDay removed.
