+++ public/app/features/dashboard/all.ts
@@ -28,7 +28,7 @@ import './move_to_folder_modal/move_to_folder';
 import './settings/settings';
 import './panellinks/module';
 import './dashlinks/module';
-import './customTimeRanges/module';
+import './custom_time_ranges/module';
 
 // angular wrappers
 import { react2AngularDirective } from 'app/core/utils/react2angular';
